+++ src/Core/Interfaces/Reposiories/IGenericRepository.cs
@@ -5,14 +5,14 @@ namespace Core.Interfaces.Reposiories
     public interface IGenericRepository<T> where T : BaseEntity
     {
         Task<IReadOnlyList<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
         Task<T> GetEntityWithSpecAsync(ISpecification<T> specification);
         Task<IReadOnlyList<T>> GetAllWithSpecAsync(ISpecification<T> specification);
         Task<int> CountAsync(ISpecification<T> specification);
         void Add(T entity);
         void Update(T entity);
         void AddRange(IReadOnlyList<T> entities);
-        void Delete(int id);
+        void Delete(long id);
         Task<int> Complete();
     }
 }
