+++ src/Core/Interfaces/Services/ICategoryService.cs
@@ -8,11 +8,11 @@ namespace Core.Interfaces.Services
     public interface ICategoryService
     {
         Task<IReadOnlyList<Category>> GetAllCategoriesAsync(CategorySpecParams specParams);
-        Task<Category> GetCategoryByIdAsync(int id);
+        Task<Category> GetCategoryByIdAsync(long id);
         Task<Category> AddCategoryAsync(CreateCategoryDto category);
         Task<IReadOnlyList<Category>> AddRangeCategoryAsync(IReadOnlyList<CreateCategoryDto> categories);
-        Task<Category> UpdateCategoryAsync(int id, CreateCategoryDto category);
-        Task<int> DeleteCategoryAsync(int id);
+        Task<Category> UpdateCategoryAsync(long id, CreateCategoryDto category);
+        Task<long> DeleteCategoryAsync(long id);
     }
 
 }
