+++ src/Core/Interfaces/Services/IOrderService.cs
@@ -4,9 +4,9 @@ namespace Core.Interfaces.Services
 {
     public interface IOrderService
     {
-        Task<Order> CreateOrderAsync(string buyerEmail, int deliveryMethod, string basketId, Address shipAddress);
+        Task<Order> CreateOrderAsync(string buyerEmail, long deliveryMethod, string basketId, Address shipAddress);
         Task<IReadOnlyList<Order>> GetOrdersByEmailAsync(string email);
-        Task<Order> GetOrderByIdAsync(int id, string email);
+        Task<Order> GetOrderByIdAsync(long id, string email);
         Task<IReadOnlyList<DeliveryMethod>> GetDeliveryMethodsAsync();
     }
 }
