+++ src/Core/Interfaces/Services/IProductBrandService.cs
@@ -8,11 +8,11 @@ namespace Core.Interfaces.Services
     public interface IProductBrandService
     {
         Task<IReadOnlyList<ProductBrand>> GetAllProductBrandsAsync(ProductBrandSpecParams specParams);
-        Task<ProductBrand> GetProductBrandByIdAsync(int id);
+        Task<ProductBrand> GetProductBrandByIdAsync(long id);
         Task<ProductBrand> AddProductBrandAsync(CreateProductBrandDto productBrandDto);
         Task<IReadOnlyList<ProductBrand>> AddRangeProductBrandAsync(IReadOnlyList<CreateProductBrandDto> productBrandDto);
-        Task<ProductBrand> UpdateProductBrandAsync(int id, CreateProductBrandDto productBrandDto);
-        Task<int> DeleteProductBrandAsync(int id);
+        Task<ProductBrand> UpdateProductBrandAsync(long id, CreateProductBrandDto productBrandDto);
+        Task<long> DeleteProductBrandAsync(long id);
         Task<int> CountAllAsync(ProductBrandSpecParams specParams);
     }
